+++ src/app/account/mod.rs
@@ -55,7 +55,6 @@ impl Account {
             | Currency::Eth
             | Currency::Gno
             | Currency::House(_)
-            | Currency::Stock(_)
             | Currency::StockPlus(_) => (
                 Transactions::new(Fiat::Usd),
                 Some(Transactions::new(currency)),